Flytte StegMedNavigasjon til nav-soknad
