--- web/src/frontend/src/digisos/skjema/konstanter.ts
@@ -1,1 +1,0 @@
-export const ANTALL_STEG = 9;
