+++ web/src/frontend/src/nav-soknad/soknadTypes.ts
@@ -109,3 +109,7 @@ export interface SkjemaSteg {
 	cmskey: string;
 	type: SkjemaStegType;
 }
+
+export interface SkjemaConfig {
+	steg: SkjemaSteg[];
+}
